refactor: clean up configuration dataclasses
