+++ src/statsvy/data/comparison_config.py
@@ -1,0 +1,10 @@
+"""Comparison configuration data model."""
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class ComparisonConfig:
+    """Comparison and delta settings."""
+
+    show_unchanged: bool
