+++ src/statsvy/data/display_config.py
@@ -1,0 +1,11 @@
+"""Display configuration data model."""
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class DisplayConfig:
+    """Terminal display preferences."""
+
+    truncate_paths: bool
+    show_percentages: bool
