+++ src/statsvy/data/storage_config.py
@@ -1,0 +1,10 @@
+"""Storage configuration data model."""
+
+from dataclasses import dataclass
+
+
+@dataclass(frozen=True, slots=True)
+class StorageConfig:
+    """Persistence settings."""
+
+    auto_save: bool
